refactor(chat): neutralize core event service filenames
