+++ turbo/apps/api/src/signals/services/chat-goal-queue.service.ts
@@ -17,11 +17,11 @@ import {
   insertChatEvent,
   revokeChatEvent,
   replaceChatEvent,
-} from "./zero-chat-event.service";
+} from "./chat-event.service";
 import { chatThreadAdmissionBlocked } from "./zero-chat-active-run.service";
-import { chatEventTypeIn } from "./zero-chat-event-type.service";
+import { chatEventTypeIn } from "./chat-event-type.service";
 import { appendGoalCloseMarker } from "./zero-chat-goal-marker.service";
-import { createUserMessageDocument } from "./zero-chat-user-message.service";
+import { createUserMessageDocument } from "./chat-user-message.service";
 import { lockGoalThread } from "./goal-lock.service";
 import {
   canonicalChatEventGoalId,
