+++ turbo/apps/api/src/signals/services/internal-feishu-chat-run-callback.service.ts
@@ -23,7 +23,7 @@ import {
 } from "./feishu-chat-callback-payload";
 import { loadUserFeatureSwitchContext } from "./feature-switches.service";
 import { resolveIntegrationAgentResponsePresentation } from "./integration-agent-response-presentation.service";
-import { chatEventTypeIn } from "./zero-chat-event-type.service";
+import { chatEventTypeIn } from "./chat-event-type.service";
 import { canonicalChatEventContent } from "./canonical-chat-event-read.service";
 
 const L = logger("InternalCallbacksFeishuChat");
