+++ turbo/apps/api/src/signals/services/morning-brief-collect.service.ts
@@ -27,8 +27,8 @@ import { loadConnectorRuntimeSnapshot } from "./connector-catalog-runtime.servic
 import {
   projectUserMessage,
   requiredUserMessageForEvent,
-} from "./zero-chat-user-message.service";
-import { chatEventTextCondition } from "./zero-chat-event-type.service";
+} from "./chat-user-message.service";
+import { chatEventTextCondition } from "./chat-event-type.service";
 import {
   canonicalChatEventContent,
   canonicalChatEventUserMessage,
