+++ turbo/apps/api/src/signals/services/zero-teams-dispatch.service.ts
@@ -70,13 +70,13 @@ import {
   publishTeamsChanged$,
 } from "./zero-teams-connect.service";
 import { touchChatThreadLastMessageAt } from "./zero-chat-event-shared.service";
-import { insertChatEvent } from "./zero-chat-event.service";
+import { insertChatEvent } from "./chat-event.service";
 import { createChatEventSourcePart } from "./chat-event-annotation.service";
-import { createUserMessageDocument } from "./zero-chat-user-message.service";
+import { createUserMessageDocument } from "./chat-user-message.service";
 import {
   chatEventTypeIn,
   chatInputPromptDispatchCondition,
-} from "./zero-chat-event-type.service";
+} from "./chat-event-type.service";
 
 const L = logger("TeamsDispatch");
 const TEAMS_LOGIN_PROMPT_FALLBACK_TEXT =
